+++ shm/src/alcohol.cpp
@@ -3,7 +3,7 @@
 Alcohol::Alcohol(const std::string& name, size_t amount, size_t basePrice)
     : Alcohol(name, amount, basePrice, kBasePower) {}
 
-Alcohol::Alcohol(const std::string& name, size_t amount, size_t basePrice, size_t power)
+Alcohol::Alcohol(const std::string& name, size_t amount, size_t basePrice, float power)
     : Cargo(name, amount, basePrice), power_(power) {}
 
 size_t Alcohol::getPrice() const {
